+++ apps/ledger-live-mobile/src/analytics/hooks/useTrackReceiveFlow.test.ts
@@ -11,11 +11,6 @@ import {
 } from "@ledgerhq/errors";
 import { CONNECTION_TYPES, HOOKS_TRACKING_LOCATIONS } from "./variables";
 
-jest.mock("../segment", () => ({
-  track: jest.fn(),
-  setAnalyticsFeatureFlagMethod: jest.fn(),
-}));
-
 describe("useTrackReceiveFlow", () => {
   const deviceMock = {
     modelId: "nanoX",