+++ e2e/desktop/tests/fixtures/common.ts
@@ -4,8 +4,7 @@ import merge from "lodash/merge";
 import * as path from "path";
 import { OptionalFeatureMap } from "@ledgerhq/types-live";
 import { getEnv, setEnv } from "@ledgerhq/live-env";
-import { startSpeculos, stopSpeculos, specs } from "@ledgerhq/live-common/e2e/speculos";
-import invariant from "invariant";
+import { stopSpeculos } from "@ledgerhq/live-common/e2e/speculos";
 
 import { Application } from "tests/page";
 import { safeAppendFile } from "tests/utils/fileUtils";
@@ -15,6 +14,9 @@ import { randomUUID } from "crypto";
 import { AppInfos } from "@ledgerhq/live-common/e2e/enum/AppInfos";
 import { lastValueFrom, Observable } from "rxjs";
 import { CLI } from "../utils/cliUtils";
+import { launchSpeculos } from "tests/utils/speculosUtils";
+
+type CliCommand = (appjsonPath: string) => Observable<unknown> | Promise<unknown> | string;
 
 type TestFixtures = {
   lang: string;
@@ -31,16 +33,23 @@ type TestFixtures = {
   featureFlags: OptionalFeatureMap;
   simulateCamera: string;
   app: Application;
-  cliCommands?: ((appjsonPath: string) => Observable<unknown> | Promise<unknown> | string)[];
+  cliCommands?: CliCommand[];
+  cliCommandsOnApp?: {
+    app: AppInfos;
+    cmd: CliCommand;
+  }[];
 };
 
 const IS_NOT_MOCK = process.env.MOCK == "0";
 const IS_DEBUG_MODE = !!process.env.PWDEBUG;
 if (IS_NOT_MOCK) setEnv("DISABLE_APP_VERSION_REQUIREMENTS", true);
 setEnv("SWAP_API_BASE", process.env.SWAP_API_BASE || "https://swap-stg.ledger-test.com/v5");
-const BASE_PORT = 30000;
-const MAX_PORT = 65535;
-let portCounter = BASE_PORT; // Counter for generating unique ports
+
+async function executeCliCommand(cmd: CliCommand, userdataDestinationPath?: string) {
+  const promise = await cmd(`${userdataDestinationPath}/app.json`);
+  const result = promise instanceof Observable ? await lastValueFrom(promise) : await promise;
+  console.log("CLI result: ", result);
+}
 
 export const test = base.extend<TestFixtures>({
   env: undefined,
@@ -52,6 +61,7 @@ export const test = base.extend<TestFixtures>({
   simulateCamera: undefined,
   speculosApp: undefined,
   cliCommands: [],
+  cliCommandsOnApp: [],
 
   app: async ({ page }, use) => {
     const app = new Application(page);
@@ -81,6 +91,7 @@ export const test = base.extend<TestFixtures>({
       simulateCamera,
       speculosApp,
       cliCommands,
+      cliCommandsOnApp,
     },
     use,
     testInfo,
@@ -95,40 +106,29 @@ export const test = base.extend<TestFixtures>({
     const userData = merge({ data: { settings } }, fileUserData);
     await fsPromises.writeFile(`${userdataDestinationPath}/app.json`, JSON.stringify(userData));
 
-    let device: any | undefined;
+    let speculos: any | undefined;
 
     try {
       if (IS_NOT_MOCK && speculosApp) {
-        // Ensure the portCounter stays within the valid port range
-        if (portCounter > MAX_PORT) {
-          portCounter = BASE_PORT;
-        }
-        const speculosPort = portCounter++;
         setEnv("PLAYWRIGHT_RUN", true);
-        setEnv(
-          "SPECULOS_PID_OFFSET",
-          (speculosPort - BASE_PORT) * 1000 + parseInt(process.env.TEST_WORKER_INDEX || "0") * 100,
-        );
-        device = await startSpeculos(
-          testInfo.title.replace(/ /g, "_"),
-          specs[speculosApp.name.replace(/ /g, "_")],
-        );
-        invariant(device, "[E2E Setup] Speculos not started");
-        invariant(device.ports.apiPort, "[E2E Setup] speculosApiPort not defined");
-        const speculosApiPort = device.ports.apiPort.toString();
-
-        setEnv("SPECULOS_API_PORT", speculosApiPort);
         setEnv("MOCK", "");
-        process.env.SPECULOS_API_PORT = speculosApiPort;
         process.env.MOCK = "";
 
+        if (cliCommandsOnApp?.length) {
+          for (const { app, cmd } of cliCommandsOnApp) {
+            speculos = await launchSpeculos(app.name);
+            CLI.registerSpeculosTransport(speculos.ports.apiPort.toString());
+            await executeCliCommand(cmd, userdataDestinationPath);
+            await stopSpeculos(speculos.id);
+          }
+        }
+
+        speculos = await launchSpeculos(speculosApp.name, testInfo.title);
+
         if (cliCommands?.length) {
-          CLI.registerSpeculosTransport(speculosApiPort);
+          CLI.registerSpeculosTransport(speculos.ports.apiPort.toString());
           for (const cmd of cliCommands) {
-            const promise = await cmd(`${userdataDestinationPath}/app.json`);
-            const result =
-              promise instanceof Observable ? await lastValueFrom(promise) : await promise;
-            console.log("CLI result: ", result);
+            await executeCliCommand(cmd, userdataDestinationPath);
           }
         }
       }
@@ -169,15 +169,14 @@ export const test = base.extend<TestFixtures>({
       // close app
       await electronApp.close();
     } finally {
-      if (device) {
-        await stopSpeculos(device.id);
+      if (speculos) {
+        await stopSpeculos(speculos.id);
       }
     }
   },
   page: async ({ electronApp }, use, testInfo) => {
     // app is ready
     const page = await electronApp.firstWindow();
-
     // we need to give enough time for the playwright app to start. when the CI is slow, 30s was apprently not enough.
     page.setDefaultTimeout(120000);