+++ libs/coin-modules/coin-sui/jest.config.js
@@ -6,4 +6,5 @@ module.exports = {
   preset: "ts-jest",
   testEnvironment: "node",
   testPathIgnorePatterns: ["lib/", "lib-es/", ".*\\.integration\\.test\\.[tj]s"],
+  setupFilesAfterEnv: ["@ledgerhq/disable-network-setup"],
 };