+++ libs/ledgerjs/packages/types-live/src/feature.ts
@@ -221,7 +221,7 @@ export type Features = CurrencyFeatures & {
   lldSolanaNfts: DefaultFeature;
   llmSolanaNfts: DefaultFeature;
   largemoverLandingpage: DefaultFeature;
-  llmMmkvMigration: DefaultFeature;
+  llmMmkvMigration: Feature_LlmMmkvMigration;
   lldModularDrawer: Feature_LldModularDrawer;
 };
 
@@ -280,27 +280,24 @@ export type Feature_SwapWalletApiPartnerList = Feature<{
   list: string[];
 }>;
 
-export type RedirectQueryParam<ManifestId> = "stakekit" extends ManifestId
+export type PlatformManifestId = "stakekit" | "kiln-widget" | "earn";
+
+export type RedirectQueryParam<M extends PlatformManifestId> = "stakekit" extends M
   ? {
       yieldId: string;
     }
-  : "kiln-widget" extends ManifestId
-    ? {
-        chaidId: number;
-      }
-    : unknown;
+  : unknown;
 
-export type Redirect<ManifestId> = {
-  platform: ManifestId;
-  /** @developer asssetId resolves to string but is either CryptoCurrency["id"] | TokenCurrency["id"]; */
-  assetId: string;
+export type Redirect<M extends PlatformManifestId> = {
+  platform: PlatformManifestId;
   name: string;
-  queryParams?: Record<string, string> & RedirectQueryParam<ManifestId>;
+  queryParams?: Record<string, string> & RedirectQueryParam<M>;
 };
 
-export type Feature_StakePrograms<ManifestId = "stakekit" | "kiln-widget"> = Feature<{
+export type Feature_StakePrograms = Feature<{
   list: string[];
-  redirects: Record<string, Redirect<ManifestId>>;
+  /** redirects is a dictionary of crypto asset ids to partner app params for overriding flows for specific tokens. */
+  redirects: Record<string, Redirect<PlatformManifestId>>;
 }>;
 
 export type Feature_StakeAccountBanner = Feature<{ [blockchainName: string]: any }>;
@@ -587,6 +584,10 @@ export type Feature_LlCounterValueGranularitiesRates = Feature<{
 
 export type Feature_LlMevProtection = Feature<{
   link: string | null;
+}>;
+
+export type Feature_LlmMmkvMigration = Feature<{
+  shouldRollback: boolean | null;
 }>;
 
 export type Feature_LldModularDrawer = Feature<{